+++ test/module/shared_model/validators/transaction_validator_test.cpp
@@ -22,7 +22,7 @@
 
 #include <type_traits>
 #include "builders/protobuf/transaction.hpp"
-#include "test_transaction_builder.hpp"
+#include "test_builder.hpp"
 #include "utils/polymorphic_wrapper.hpp"
 
 using namespace shared_model;
